requirements.txt
anyio==3.6.2
boto3==1.26.64
botocore==1.29.64
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
jmespath==1.0.1
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
packaging==23.0
Pillow==9.4.0
pony==0.7.16
psycopg2-binary==2.9.5
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
python-telegram-bot==20.0
PyYAML==6.0
randimage==1.2
requests==2.28.2
rfc3986==1.5.0
s3transfer==0.6.0
scipy==1.10.0
six==1.16.0
sniffio==1.3.0
urllib3==1.26.14
